Add `ViewerContext::save_file_dialog`
