+++ crates/viewer/re_viewer_context/Cargo.toml
@@ -57,6 +57,7 @@ ndarray.workspace = true
 nohash-hasher.workspace = true
 once_cell.workspace = true
 parking_lot = { workspace = true, features = ["serde"] }
+rfd.workspace = true
 serde.workspace = true
 slotmap.workspace = true
 smallvec.workspace = true
